container: Move to its own crate
Our container implementation could be used externally.
Moreover, and in order to have a cleaner separation between the kaps CLI
and binary and the container API, we should move the container code into
its own crate.

With that move, kaps becomes a simple CLI consuming the container crate.

Fixes #17
