+++ src/cli/run.rs
@@ -1,6 +1,6 @@
-use crate::container::Container;
 use crate::{Handler, Result};
 use clap::Args;
+use container::Container;
 
 /// Arguments for our `RunCommand`.
 ///
