+++ src/main.rs
@@ -3,7 +3,6 @@ use clap::Parser;
 use crate::cli::{Cli, Handler, Result};
 
 mod cli;
-mod container;
 
 fn main() -> Result<()> {
     let cli: Cli = Cli::parse();
